requirements.txt
asgiref==3.7.2
Django==3.2.10
djangorestframework==3.14.0
pytz==2024.1
sqlparse==0.4.4
tzdata==2024.1
django-environ==0.11.2  
